Start server only when we see *.lean files.
Previously the server was started during initialization of the
extension.  Since we now enable the lean input mode also in markdown
files, the extension gets loaded (and a server is started) even if there
are no lean files present.

Additionally, we spawned a second lean process to determine the lean
version.  For elan users, this caused two elan instances to download
lean concurrently since elan doesn't use locks.

As a side effect, we drop compatibility for lean versions older than
3.1.0.