fix range in diagnostics:
Previously, if the position of a message from the server
happened to lie between "words" (e.g. at the end of a file),
getWordRangeAtPosition would return undefined, leading
to a broken range. This change adds a fallback to the old
behavior of creating a Range that both begins and ends
at the position of the server message.