Remove red code marker when hovering over infoview
It should be easy to add features where the infoview can either reveal
or highlight positions in the sourcetext using the 'reveal' and
'hover_position' and 'stop_hover' `FromInfoViewMessage` commands.
But how they manifest currently is a bit garish and confusing so I will
wait for someone with some design sense to figure out what is best for a
good user experience.