Add a new event model for updating widgets.
This means that getInfo tasks are now marshalled so that only one info request is in flight at a given time.
Getting info is fast enough that this might as well be enabled all the time.
This hopefully squashes a large category of bugs where multiple inflight getInfos would mess up state.