Fix issue where it is showing no infos found when it needs to update.
It happens when the tactic completes too fast and the user types quickly, and so when
lean reports that the task is done, there was no intermediate task reported
so occasionally you would still get no info found bugs.

Right now it is quite jumpy and will flash 'not found' messages on the screen
The long term fix for this is going to be adding a flag to the lean info response saying whether there is no
tactic state or widget because it is loading or because there is nothing there.