Add support for katex.
The design is that now if the user uses the special tags `InlineMath` or `BlockMath`, as in;
```
```
Then vscode will catch it and turn it in to a component. The component list is stored in a javascript object called `tagComponentDict`.
This does make the lean <-> vscode interface more complicated, because now lean is assuming that the widget renderer knows to render InlineMath and BlockMath as katex components.
So for example if someone wanted to write thier own widget renderer, there are now more assumptions that they need to be aware of.
However in the case of some key libraries like D3 and Katex I think this is not too great a sacrifice since the only widgets client at the moment is the vscode extension.

It would be nice at some point to have Lean be able to programattically say which libraries it wants, but for now this will do.
Alternatively, at some point we should standardise which libraries and CSS classes lean expects to be available.