Escape "-" in wordPattern regexp
Since "-" was not escaped, the sequence ")-=" was treated as "a range from ) to =". 

Escaping "-" fixes this.

Examples:

```
const r = new RegExp("([^`~@$%^&*()\\-=+\\[{\\]}⟨⟩⦃⦄⟦⟧⟮⟯‹›\\\\|;:\",./\\s]+)")

'SomeName123'.match(r)[0] === 'SomeName123'
'Some-Name-123'.match(r)[0] === 'Some'
```

P.S. Two backslashes are needed because the "wordPattern" is a key in a JSON file, which requires its own escape sequence.