feat: support for multi-toolchain workspaces (#428)
This PR adds support for using multiple lean toolchains in the same
workspace folder. It also enables users to use Lean 4 without first
having to open a folder.

For nested workspaces, the client uses the outer-most Lean project. This
is the correct behavior for projects in `.lake`, but not the correct
behavior for actual nested projects. I unfortunately do not see a way to
make the latter work because the language client library only supports
globs to specify which files should be picked up by a language server,
and hence cannot express something like "all Lean files in this
directory, except those contained in inner directories with their own
lean-toolchain". We could in principle set up middleware for every
single notification and request, but this is likely to be overly
fragile.

Closes #138.