feat: better progress ux (#542)
This PR improves two things about the way we report progress:
- Instead of attempting to badly approximate an accurate progress bar
that gets stuck whenever the output of a command doesn't change, we now
use VS Code's "infinite" progress bars everywhere that clearly show that
an operation is actively running in the background.
- Every progress bar is prefixed with a context to make it clear which
operation caused the operation. For example, when creating a new mathlib
project, all external commands executed while creating the project are
prefixed with "[Create Mathlib Project]".

Closes #457. IMO, our dialog UX still isn't perfect, but there doesn't
seem to be a good way to improve these things short of implementing our
own webviews for every command.