feat: support for elan 4.0.0 (#554)
This PR implements support for Elan 4.0.0. Specifically:
- A new setting that can be enabled to issue confirmation prompts before
installing new toolchains
- Confirmation prompts before eagerly updating release channels when
opening Lean files in VS Code with Elan 4.0.0's eager toolchain
resolution
- New VS Code command integrations for Elan:
- Project and default toolchain selection that displays all available
Lean toolchains by using https://release.lean-lang.org/
- Toolchain deinstallation, including integration for `elan toolchain
gc` in Elan 4.0.0
  - Manual updates of release channel toolchains

It also fixes a couple of bugs:
- When the output of `elan show` got too large, the 'Troubleshooting:
Show Setup Information' command would display a dialog that could exceed
the height of the screen, hiding all buttons. With Elan 4.0.0, the
content of this dialog is now kept much smaller so that this cannot
happen on most screens anymore.
- The InfoView would sometimes report internal errors (including
stacktraces) when something went wrong on the VS Code side of the
InfoView. Now it correctly displays the error.
- The order of the buttons in lots of modal dialogs was a bit
unorganized, and it should now be more coherent.
- If the default toolchain was changed while VS Code was running, the VS
Code extension would hold on to the old toolchain for the language
client of untitled files indefinitely.
- VS Code extension commands that use the active client of the client
provider would behave incorrectly or use an incorrect client if a client
failed or was stopped manually using a command.
- The 'Restart Server' command didn't always restart the setup check for
the selected language client if that language client had failed in the
past.

<details> 
  <summary>Update dialog</summary>
  

![Update_Dialog](https://github.com/user-attachments/assets/8d521416-acb6-44aa-b6e0-1acee38c5f58)
</details>

<details> 
  <summary>Installation dialog</summary>
  

![Installation_Dialog](https://github.com/user-attachments/assets/238d03fa-e227-4d78-b04e-bb95244c5c0d)
</details>

<details> 
  <summary>New version management commands</summary>


![New_Version_Management_Menu](https://github.com/user-attachments/assets/0dd8ccc1-c59e-4e15-8dda-e3671fba5bf3)
</details>

<details> 
  <summary>Default version dialog</summary>


![Default_Version_Dialog](https://github.com/user-attachments/assets/8cd6dc4d-8b6d-4653-8e46-ab9fa83e8e96)
</details>

<details> 
  <summary>Release channel update dialog</summary>


![Release_Channel_Update_Dialog](https://github.com/user-attachments/assets/6d687465-0035-47df-b7f6-9a55a70597f6)
</details>

<details> 
  <summary>Version uninstall dialog</summary>


![Version_Uninstall_Dialog](https://github.com/user-attachments/assets/029a64ca-c40a-4704-a570-c209edffdc1f)
</details>

This PR will be merged at the start of 2025.