feat: improve activation error when forked extension is installed (#613)
Every so often, users decide that they want to get (what they believe to
be) the "full" Lean experience in VS Code and install every single
Lean-related extension from the VS Code marketplace. While
well-intentioned, since the marketplace also contains forks of our
extension, this then leads to an internal activation error when VS
Code's `commands.registerCommand` API throws an exception for trying to
register duplicate commands.
This PR improves the error message that is displayed in this case to
hint at the cause so that users are more likely to be able to fix it
themselves without requiring assistance.