fix: remove the redundant `include: #stringBlock` in lean4 syntax
`#stringBlock` was added in 48ad018 on 2020-06-23, when people were developing in the legacy vscode-lean repo.
Later, `#stringBlock` was removed in ca65431 on 2021-01-18 when migrating from lean 3 to lean 4. However, the `include` line was forgotten.

In fact, VS Code’s parser (vscode-textmate) will tolerate missing includes, so the original syntax works.