fix: offer update options for all dependencies in manifest (#638)
This PR removes the logic from the 'Update Dependency' command that
hides dependencies that are already up-to-date relative to the manifest
and removes the revision information from the selection dialog.

The reason for this is that the manifest is not an up-to-date
representation of the set of dependencies in the `lakefile.lean` or the
`lakefile.toml`, e.g. if the input revision in the Lake configuration
file was changed. I don't know of a way to query this information from
Lake without also updating all the revisions of the manifest in the
process, so we now just avoid this situation entirely by considering the
revision information in the manifest to be unreliable.

Note that there is still a shortcoming of this approach: When adding a
new dependency to the `lakefile.lean` or the `lakefile.toml`, we won't
display it in the update dialog until the manifest is updated.

This came up at [#general > Updating Mathlib via VSCode menu does not
work @
💬](https://leanprover.zulipchat.com/#narrow/channel/113488-general/topic/Updating.20Mathlib.20via.20VSCode.20menu.20does.20not.20work/near/526969295).