feat: better installation (#641)
This PR reworks the installation mechanism of Lean in the VS Code
extension. Git and curl are now installed automatically, without users
having to open the terminal. The precondition check that checks for Git
and curl also now offers to install them for the user. The 'Install
Dependencies' and 'Install Elan' steps of the setup guide are now one
single 'Install Lean' step.

On Windows, `winget` is used if available. Otherwise, we download a
Git-for-Windows installer and run it silently.

On macOS, Apple Command Line Tools is installed for Git.

On Linux systems that don't support `pkexec`, a manual installation
procedure is offered. For systems that don't use `apt-get` or `dnf`, the
mechanism relies on the user to know what they are doing and suggests to
install `git` and `curl` without offering an installation script.
Otherwise, `git` and `curl` are installed using either `apt-get` or
`dnf`.

### Other changes
- The whole installation procedure is now using sub-processes, not the
built-in terminal of VS Code.
- The Windows installation procedure is now using `Invoke-WebRequest`.
- The documentation page of the setup guide has been updated to make
similar suggestions as the lean-lang website.
- The config watch service (the mechanism that displays a notification
when you changed your lakefile or lean-toolchain) is replaced with a
mechanism that hopefully causes fewer false-positives.

Closes #349, #558.