feat: increase delay until completion requests are emitted (#654)
This PR increases the delay until VS Code emits an auto-completion
request from 10ms to 200ms. A delay of 10ms means that VS Code will
immediately emit a completion request for the first character for most
users, even if they continue typing, which tends to be very expensive if
e.g. `import Mathlib` is at the top of the file.