feat: update textmate grammar (#682)
This PR adds newer keywords that have been added to Lean 4 and removes
ones that have been removed from Lean 4 in the mean-time.

It also mitigates #609 by limiting the situations where character
highlighting occurs and mitigates #515 by not attempting to highlight
the brackets (and thus needing to find a closing bracket) when the
attribute declaration has parameters. Given the nature of the TextMate
grammar, neither of these are complete fixes.

#369 and #656 remain problematic for now. In the future, we are planning
to replace the markdown highlighting in the TextMate grammar with
semantic highlighting in the language server.

Closes #609, closes #515.