fix: remove backtick name highlighting (#692)
This PR removes the text mate grammar scope for name literals. As far as
I can tell, it isn't being used in the VS Code default themes.

Example of something that is causing problems with the current
highlighting where the whole file is highlighted:
```
def x := s!"{`False} "

def y := 1
```