test: add tests for `lakefile-toml-schema.json` (#694)
This pull request adds tests to make sure that
[lakefile-toml-schema](https://raw.githubusercontent.com/leanprover/lean4/refs/heads/master/src/lake/schemas/lakefile-toml-schema.json)
has no false positives and minimal false negatives *(or the other way
around? Maintainer please confirm/deny)*.

As discussed in [this pull
request](https://github.com/leanprover/lean4/pull/11536), tests for the
`lakefile.toml` schema are placed here since the tests in this repo are
better maintained and this location avoids adding another dependency to
the test suite for the main project.